requirements.txt
flask==3.0.0
flask-cors==4.0.0
bcrypt==4.1.2
google-generativeai==0.3.1
